sorted out GLFWScreen
